+++ clients/typescript/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2025-01-21
+
+### Changed
+
+- Shared mode for the `server` command has been replaced by a flag to
+  specify a container name. If the container name is set, the server
+  uses the preexisting container with that name. The regular `up` and
+  `down` commands can be used to manage that container.
+
 ## [0.5.0] - 2025-01-20
 
 ### Added